% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ews.R
\name{calculate_EWS}
\alias{calculate_EWS}
\title{Calculate the early warning signals}
\usage{
calculate_EWS(dataIn, winMove)
}
\arguments{
\item{dataIn}{Used in calc_FisherInformation. Default = 2 data points}

\item{winMove}{Proportion of data to be included in each moving window (0,1).}

\item{distances}{A data frame of the distances and dervatives of distance travelled at each time point.}
}
