% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fisherInformation.R
\name{calculate_FisherInformation}
\alias{calculate_FisherInformation}
\title{Fisher Information: three equations for calculating.}
\usage{
calculate_FisherInformation(dataInFI, min.window.dat = 2,
  fi.equation = "7.12")
}
\arguments{
\item{dataInFI}{A subset of data for each moving window. These data will be used to calculate the Fisher Information}

\item{min.window.dat}{The minimum number of observation points (e.g. time) within the moving window for Fisher Information to be calculated}

\item{fi.equation}{Equations descirbed in Mayer et al. (2007)}
}
\description{
Fisher Information: three equations for calculating.
}
\references{
Mayer, Audrey L., et al. "Applications of Fisher information to the management of sustainable environmental systems." Exploratory data analysis using Fisher information. Springer, London, 2007. 217-244.
}
