% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_varianceIndex.R
\name{calculate_VI}
\alias{calculate_VI}
\title{Calculate the Variance Index}
\usage{
calculate_VI(winData)
}
\description{
Calculate the Variance Index
}
\references{
[1] Brock and Carpenter YYYY
}
