% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_results.R
\name{interpolate_distance}
\alias{interpolate_distance}
\title{Linearally Interpolate the Results for Distance Travelled}
\usage{
interpolate_distance(data, timeOut)
}
\arguments{
\item{data}{A data frame containing results from `calculate_distanceTravelled()`}

\item{timeOut}{The number of time units over which we will interpolate (with equal spacing).}
}
\description{
Linearally Interpolate the Results for Distance Travelled
}
