% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_results.R
\name{interpolate_FI_VI}
\alias{interpolate_FI_VI}
\title{Linearally Interpolate the Results for Distance Travelled}
\usage{
interpolate_FI_VI(EWS, timeOut, winMove)
}
\arguments{
\item{timeOut}{The number of time units over which we will interpolate (with equal spacing).}

\item{data}{A data frame containing results from `calculate_distanceTravelled()`}
}
\description{
Linearally Interpolate the Results for Distance Travelled
}
