% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{munge_orig_dat}
\alias{munge_orig_dat}
\title{Munge the Original Data Frame.}
\usage{
munge_orig_dat(data = NULL, example = T, fill = NA)
}
\arguments{
\item{data}{A data frame with ENTER DESCRIPTION}

\item{example}{Loads and munges data when parameters data = NULL & Default = T, running the function will load and munge data from [1].}

\item{fill}{Fills empty cells with this value. Default = NA. Consider using zero (0) as a fill if using species counts. Beware when using NA vs. zero.}
}
\value{
Function returns a data frame in long format with columns specifying site name, time (or spatial unit), variable (e.g. species identity), and value (e.g. species count).
}
\examples{
munge_orig_dat() # will not write an object to environment
newDf = munge_orig_dat() # save the object to environment
}
\references{
[1] Spanbauer, Trisha L., et al. "Prolonged instability prior to a regime shift." PLoS One 9.10 (2014): e108936.
}
