% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_originalData.R
\name{plot_timeDiff}
\alias{plot_timeDiff}
\title{Plot units of time elapsed between sampling points over the time series.
This function plots the amount of time that has elapsed since the last sampled unit time.}
\usage{
plot_timeDiff(data, example = FALSE, print = TRUE, save = FALSE,
  xLabel = "time")
}
\arguments{
\item{example}{Logical. If TRUE will use the paleodiatom data from [1], else will use input data.}

\item{print}{print plots to device when print = T. Default print = T.}

\item{save}{UNDER DEV: NEEDS TO BE UPDATED TO CREATE LOCAL PLOT FOLDER AND EXPORT TO THIS FOLDER. JLB.}

\item{origData}{The original data frame}
}
\value{
Function returns a data frame in long format with columns specifying site name, time (or spatial unit), variable (e.g. species identity), and value (e.g. species count).
}
\description{
Plot units of time elapsed between sampling points over the time series.
This function plots the amount of time that has elapsed since the last sampled unit time.
}
\examples{
df <- munge_orig_dat()
plot_timeDiff(df)

}
\references{
[1] Spanbauer, Trisha L., et al. "Prolonged instability prior to a regime shift." PLoS One 9.10 (2014): e108936.
}
