% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distanceTravelled.R
\name{calculate_distanceTravelled}
\alias{calculate_distanceTravelled}
\title{Calculate the `distance travelled` by the entire system. Also calculates the velocity and acceleration of the entire system over the time series.}
\usage{
calculate_distanceTravelled(dataIn, derivs = T, print = T)
}
\arguments{
\item{print}{}
}
