% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fisherInformation.R
\name{calculate_FisherInformation}
\alias{calculate_FisherInformation}
\title{Fisher Information: three equations for calculating.}
\usage{
calculate_FisherInformation(dataInFI, min.window.dat = 2,
  fi.equation = "7.12")
}
\arguments{
\item{dataInFI}{A subset of data for each moving window. These data will be used to calculate the Fisher Information}
}
\description{
Fisher Information: three equations for calculating.
}
