% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_varianceIndex.R
\name{calculate_VI}
\alias{calculate_VI}
\title{Calculate the Variance Index}
\usage{
calculate_VI(winData, fill = 0)
}
\arguments{
\item{winData}{A data frame.}

\item{fill}{Fill for missing data. Default = 0.}
}
\description{
The Variance Index is the maximum eigenvalue of the variance covariance matrix, and is described in Brock et al. (2006).
}
\references{
Brock, William A., and Stephen R. Carpenter. "Variance as a leading indicator of regime shift in ecosystem services." Ecology and Society 11.2 (2006).
}
