% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_originalData.R
\name{plot_orig_data}
\alias{plot_orig_data}
\title{Plot original data}
\usage{
plot_orig_data(data, example = F, print = T, save = F)
}
\arguments{
\item{print}{print plots to device when print = T. Default print = T.}

\item{save}{NEEDS TO BE UPDATED TO CREATE LOCAL PLOT FOLDER AND EXPORT TO THIS FOLDER. JLB.}
}
\value{
Function returns a data frame in long format with columns specifying site name, time (or spatial unit), variable (e.g. species identity), and value (e.g. species count).
}
\description{
Plot original data
}
\examples{
df <- munge_orig_dat()
plot_orig_data(df)

}
\references{
[1] Spanbauer, Trisha L., et al. "Prolonged instability prior to a regime shift." PLoS One 9.10 (2014): e108936.
}
