% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_originalData.R
\name{plot_richness}
\alias{plot_richness}
\title{Plot species richness over time for the original data
This function plots species richness at each time unit. Here, species richness is the total number of unique variables with unit measurement (e.g., abundance) > 0 and not NA.}
\usage{
plot_richness(data, example = F, print = T, save = F,
  xLabel = "time")
}
\arguments{
\item{print}{print plots to device when print = T. Default print = T.}

\item{save}{NEEDS TO BE UPDATED TO CREATE LOCAL PLOT FOLDER AND EXPORT TO THIS FOLDER. JLB.}
}
\value{
Function returns a data frame in long format with columns specifying site name, time (or spatial unit), variable (e.g. species identity), and value (e.g. species count).
}
\description{
This function plots the species richness over time as a function of the total unique variables within each time
}
\examples{
df <- munge_orig_dat()
plot_richness(df)

}
\references{
[1] Spanbauer, Trisha L., et al. "Prolonged instability prior to a regime shift." PLoS One 9.10 (2014): e108936.
}
