% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdm_window_analysis.R
\name{rdm_window_analysis}
\alias{rdm_window_analysis}
\title{Plot original data}
\usage{
rdm_window_analysis(dataIn, winMove = 0.25, overrideSiteErr = F,
  min.window.dat = 2, fi.equation = "7.12", to.calc = c("VI", "FI",
  "EWS"))
}
\arguments{
\item{winMove}{Number as proportion of each time series to be included in the moving windows. Default = 0.25 (25% of data included in each window).}

\item{min.window.dat}{Used in calc_FisherInformation. Default = 2 data points}

\item{to.calc}{Which measures to calculate. VI variance index. FI Fisher Information. EWS 1st through 4th moments, etc. Default = ALL measures.}

\item{min.windowdat}{Minimum # of data points in each window to include in calculations. Default = 2.}
}
